Set featured ML image ID
